move function for getting today’s galleries opening hours to common
